moved adapters to new package
