+++ WiseTrack/app/src/main/java/com/faanggang/wisetrack/adapters/CommentItemView.java
@@ -1,4 +1,4 @@
-package com.faanggang.wisetrack.comment;
+package com.faanggang.wisetrack.adapters;
 
 import android.content.Context;
 import android.view.View;
